* split analysis code into files. This is a first step towards   code refactoring
