+++ HTTAnalysis/Makefile.am
@@ -17,7 +17,8 @@ noinst_PROGRAMS=test
 
 # Sources for the a.out 
 test_SOURCES= HTTAnalysis.cc EventProxyHTT.cc HTTWeightsMaker.cc HTTWeightHistograms.cc \
-	      HTTAnalyzer.cc HTTHistograms.cc HTTEvent.cxx DictCint.cxx
+	      HTTAnalyzer.cc HTTAnalyzerCPmethods.cc HTTAnalyzerUtilityMethods.cc \
+              HTTHistograms.cc HTTEvent.cxx DictCint.cxx
 
 
 # Linker options for a.out
